Moved everything related to cart to a folder named cart under /modules
