+++ src/modules/cart/Slogan.jsx
